[BUG FIX] - UI freeze issue while running server task
[CAUSE]
- For some reason the UI thread is blocked during execution
server task in a Callable implementation in an ExecutorService.
The idea was to set a timeout when communicating with the server,
therefore, when the timeout is reached, an exception is thrown and
the flow is to interrupt. However, in this scenario, the UI got
frozen until task completed. Even running these tasks in another
thread, the UI thread waits for the child thread to return.

[SOLUTION]
 - Change Callable to Runnable implementation and don't use
Executor#awaitTermination()